+++ creator-node/src/components/replicaSet/replicaSetController.js
@@ -13,9 +13,7 @@ const {
 } = require('./URSMRegistrationComponentService')
 const { ensureStorageMiddleware } = require('../../middlewares')
 const { enqueueSync } = require('./syncQueueComponentService')
-const {
-  secondarySyncFromPrimary
-} = require('../../services/sync/secondarySyncFromPrimary')
+const secondarySyncFromPrimary = require('../../services/sync/secondarySyncFromPrimary')
 
 const router = express.Router()
